% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{addClusterMarkers2CatDisplay}
\alias{addClusterMarkers2CatDisplay}
\title{A method}
\usage{
addClusterMarkers2CatDisplay(obj, addCorCatsToLabDb = TRUE)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
