% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{addDf2seuratMetaData}
\alias{addDf2seuratMetaData}
\title{A method}
\usage{
addDf2seuratMetaData(obj, dfAdd)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
