% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{createNormSampleList}
\alias{createNormSampleList}
\title{A method}
\usage{
createNormSampleList(
  obj,
  reduce = NULL,
  vars.to.regress = NULL,
  g2m.genes = NULL,
  s.genes = NULL,
  annotateCellCyclePhase = TRUE,
  integrationReduction = "cca",
  feature_type = "Gene Expression"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
