% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{createSampleListQC}
\alias{createSampleListQC}
\title{A method}
\usage{
createSampleListQC(
  obj,
  reduce = NULL,
  vars.to.regress = NULL,
  s.genes = NULL,
  g2m.genes = NULL,
  annotateCellCyclePhase = TRUE,
  feature_type = "Gene Expression"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
