% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{createXLSXoutput}
\alias{createXLSXoutput}
\title{A method}
\usage{
createXLSXoutput(
  dfTable = "dfTable",
  outPutFN = "path/to/output/FN.xlsx",
  tableName = "Table1"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
