% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doCRplots}
\alias{doCRplots}
\title{A method}
\usage{
doCRplots(
  obj,
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4,
  dotsize = 0.5
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
