% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doDF_plotSL}
\alias{doDF_plotSL}
\title{A method}
\usage{
doDF_plotSL(
  SampleList,
  obj = "Obio",
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4,
  dotsize = 0.5
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
