% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doDGEsc}
\alias{doDGEsc}
\title{A method}
\usage{
doDGEsc(
  obj,
  DGEselCol = "sub_clusters_ExNeurons",
  colName = "DGE_name",
  contrastTag = "contrast_1_",
  DGEsampleList = list(M1 = c(5), M2 = c(1))
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
