% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doPercMT_plotSL}
\alias{doPercMT_plotSL}
\title{A method}
\usage{
doPercMT_plotSL(
  SampleList,
  obj,
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
