% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doRNAfeat_plotSL}
\alias{doRNAfeat_plotSL}
\title{A method}
\usage{
doRNAfeat_plotSL(
  SampleList,
  obj = "Obio",
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
