% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doUMAP_cellCyle}
\alias{doUMAP_cellCyle}
\title{A method}
\usage{
doUMAP_cellCyle(
  SampleList,
  obj = Obio,
  figureCount = figureCount,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4,
  dotsize = 0.5,
  s.genes = NULL,
  g2m.genes = NULL,
  cellCycleRefFile = NULL
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
