% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doUMAP_plot_percMT}
\alias{doUMAP_plot_percMT}
\title{A method}
\usage{
doUMAP_plot_percMT(
  SampleList,
  obj = "Obio",
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4,
  dotsize = 0.5
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
