% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicToolsSC.package.r
\name{doCellCycleBarchart}
\alias{doCellCycleBarchart}
\title{A method}
\usage{
doCellCycleBarchart(
  SampleList,
  obj = "Obio",
  figureCount = 1,
  VersionPdfExt = ".pdf",
  tocSubLevel = 4,
  s.genes = NULL,
  g2m.genes = NULL,
 
    cellCycleRefFile = "paste0(hpc.mount, \\"Projects/reference_data/cell_cycle_vignette_files/nestorawa_forcellcycle_expressionMatrix.txt\\")"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
